separate out integration test into utils
