+++ e2e_example/dart_test.yaml
@@ -1,1 +1,4 @@
 timeout: 2x
+# These test suites must not run in parallel, they modify actual sources in the
+# package.
+concurrency: 1
